+++ packages/core/src/sections/contact2.tsx
@@ -19,6 +19,7 @@ interface IContact2Props extends ISectionProps, IFormProps {
   // TODO(krishan711): think about how the other input fields should be customized.
   inputButtonText?: string;
   inputSuccessMessageText?: string;
+  isPhoneNumberOptional?: boolean;
 }
 
 export const Contact2 = (props: IContact2Props): React.ReactElement => {
@@ -76,7 +77,7 @@ export const Contact2 = (props: IContact2Props): React.ReactElement => {
       setErrorMessage(emailValidationResult.errorMessage || 'Email is not valid.');
       return;
     }
-    if (!phoneNumber) {
+    if (!props.isPhoneNumberOptional && !phoneNumber) {
       setPhoneNumberError('Please enter your phone number');
       return;
     }