+++ packages/core/src/application/headContent.tsx
@@ -15,14 +15,16 @@ export const HeadContent = (props: IHeadContentProps): React.ReactElement => {
   const website = useWebsite(props.website);
   const theme = mergeTheme(useTheme(), props.theme);
   const title = website.title || `${website.name} - ${website.tagline}`;
+  const assetPrefix = website.buildHash ? `/${website.buildHash}` : '';
   let socialCardImageUrl = website.socialCardImageUrl;
-  if (!socialCardImageUrl) {
-    // TODO(krishan711): set a default image!
-  } else if (socialCardImageUrl.startsWith('/')) {
-    socialCardImageUrl = `https://${website.siteHost}${socialCardImageUrl}`;
+  if (socialCardImageUrl && socialCardImageUrl.startsWith('/')) {
+    socialCardImageUrl = `${assetPrefix}${socialCardImageUrl}`;
+  }
+  let faviconImageUrl = website.faviconImageUrl;
+  if (faviconImageUrl && faviconImageUrl.startsWith('/')) {
+    faviconImageUrl = `${assetPrefix}${faviconImageUrl}`;
   }
 
-  const assetPrefix = website.buildHash ? `/${website.buildHash}` : '';
   return (
     <Head>
       <meta charSet='utf-8' />
@@ -58,14 +60,14 @@ export const HeadContent = (props: IHeadContentProps): React.ReactElement => {
       <meta property='og:title' content={title} />
       <meta property='og:site_name' content={website.name} />
       { website.description && <meta property='og:description' content={website.description} /> }
-      <meta property='og:image' content={socialCardImageUrl} />
+      {socialCardImageUrl && <meta property='og:image' content={socialCardImageUrl} />}
       <meta property='og:url' content={`https://${website.siteHost}`} />
 
       {/* Twitter */}
       <meta name='twitter:card' content={(website.iosAppId || website.androidAppId) ? 'app' : 'summary_large_image'} />
       <meta name='twitter:title' content={title} />
       { website.description && <meta name='twitter:description' content={website.description} /> }
-      <meta name='twitter:image' content={socialCardImageUrl} />
+      {socialCardImageUrl && <meta name='twitter:image' content={socialCardImageUrl} />}
       {/* <meta name='twitter:image:alt' content='Alt text for image' /> */}
       { website.twitterUsername && <meta name='twitter:site' content={`@${website.twitterUsername}`} /> }
       { website.twitterCompanyUsername && <meta name='twitter:creator' content={`@${website.twitterCompanyUsername}`} /> }
@@ -81,8 +83,8 @@ export const HeadContent = (props: IHeadContentProps): React.ReactElement => {
       <meta name='apple-mobile-web-app-title' content={website.name} />
       <meta name='apple-mobile-web-app-capable' content='yes' />
       <meta name='apple-mobile-web-app-status-bar-style' content='default' />
-      <link rel='apple-touch-icon' sizes='180x180' href={`${assetPrefix}/assets/_generated/favicon-180.png`} />
-      <link rel='apple-touch-icon' sizes='1024x1024' href={`${assetPrefix}/assets/_generated/favicon-1024.png`} />
+      {website.shouldUseGeneratedFavicons && <link rel='apple-touch-icon' sizes='180x180' href={`${assetPrefix}/assets/_generated/favicon-180.png`} />}
+      {website.shouldUseGeneratedFavicons && <link rel='apple-touch-icon' sizes='1024x1024' href={`${assetPrefix}/assets/_generated/favicon-1024.png`} />}
 
       {/* iOS Banner */}
       { website.iosAppId && <meta name='apple-itunes-app' content={website.iosAppId} /> }
@@ -90,7 +92,7 @@ export const HeadContent = (props: IHeadContentProps): React.ReactElement => {
       {/* Windows */}
       <meta name='msapplication-navbutton-color' content={String(theme.colors.brandPrimary)} />
       <meta name='msapplication-TileColor' content={String(theme.colors.brandPrimary)} />
-      <meta name='msapplication-TileImage' content={`${assetPrefix}/assets/_generated/favicon-144.png`} />
+      {website.shouldUseGeneratedFavicons && <meta name='msapplication-TileImage' content={`${assetPrefix}/assets/_generated/favicon-144.png`} />}
       <meta name='msapplication-config' content='browserconfig.xml' />
       <meta name='msapplication-tooltip' content={website.tagline || website.description} />
       <meta name='msapplication-starturl' content='/' />
@@ -101,18 +103,19 @@ export const HeadContent = (props: IHeadContentProps): React.ReactElement => {
       <meta name='browsermode' content='application' />
 
       {/* Favicons */}
-      <link rel='icon' type='image/png' sizes='16x16' href={`${assetPrefix}/assets/_generated/favicon-16.png`} />
-      <link rel='icon' type='image/png' sizes='32x32' href={`${assetPrefix}/assets/_generated/favicon-32.png`} />
-      <link rel='icon' type='image/png' sizes='48x48' href={`${assetPrefix}/assets/_generated/favicon-48.png`} />
-      <link rel='icon' type='image/png' sizes='228x228' href={`${assetPrefix}/assets/_generated/favicon-228.png`} />
-      <link rel='shortcut icon' type='image/x-icon' href={`${assetPrefix}/assets/_generated/favicon.ico`} />
+      {!website.shouldUseGeneratedFavicons && faviconImageUrl && <link rel='icon' href={faviconImageUrl} />}
+      {website.shouldUseGeneratedFavicons && <link rel='icon' type='image/png' sizes='16x16' href={`${assetPrefix}/assets/_generated/favicon-16.png`} />}
+      {website.shouldUseGeneratedFavicons && <link rel='icon' type='image/png' sizes='32x32' href={`${assetPrefix}/assets/_generated/favicon-32.png`} />}
+      {website.shouldUseGeneratedFavicons && <link rel='icon' type='image/png' sizes='48x48' href={`${assetPrefix}/assets/_generated/favicon-48.png`} />}
+      {website.shouldUseGeneratedFavicons && <link rel='icon' type='image/png' sizes='228x228' href={`${assetPrefix}/assets/_generated/favicon-228.png`} />}
+      {website.shouldUseGeneratedFavicons && <link rel='shortcut icon' type='image/x-icon' href={`${assetPrefix}/assets/_generated/favicon.ico`} />}
 
       {/* Favicons - Android */}
-      <link rel='icon' sizes='192x192' href={`${assetPrefix}/assets/_generated/favicon-192.png`} />
-      <link rel='icon' sizes='128x128' href={`${assetPrefix}/assets/_generated/favicon-128.png`} />
+      {website.shouldUseGeneratedFavicons && <link rel='icon' sizes='192x192' href={`${assetPrefix}/assets/_generated/favicon-192.png`} />}
+      {website.shouldUseGeneratedFavicons && <link rel='icon' sizes='128x128' href={`${assetPrefix}/assets/_generated/favicon-128.png`} />}
 
       {/* PWA Manifest */}
-      <link rel='manifest' href={`${assetPrefix}/assets/_generated/manifest.json`} />
+      {website.shouldUseGeneratedFavicons && <link rel='manifest' href={`${assetPrefix}/assets/_generated/manifest.json`} />}
     </Head>
   );
 };