+++ CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [Unreleased]
 
 ### Added
+- [MINOR] Add `isPhoneNumberOptional` to `Contact1` and `Contact2`
 
 ### Changed
+- [MINOR] Update `Calendly1` to hide cookie banner
 
 ### Removed